Synes det er litt kamuflert med mange modellklassar i samme fil, så deler opp i fleire
